backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
marshmallow==3.21.1
python-dotenv==1.0.1
gunicorn==22.0.0
